+++ modules/core/src/utils/flatten.js
@@ -142,85 +142,3 @@ export function flattenTypedVertices(nestedArray, {
   return start;
 }
 */
-
-/*
- * Helper function for padArray
- */
-function padSubArray({source, target, start = 0, end, enter}) {
-  end = end || target.length;
-
-  const sourceLength = source.length;
-  const targetLength = end - start;
-
-  if (sourceLength > targetLength) {
-    target.set(source.subarray(0, targetLength), start);
-    return;
-  }
-
-  target.set(source, start);
-
-  // source is not large enough to fill target space, call `enter` to get filler data
-  let i = sourceLength;
-  while (i < targetLength) {
-    const datum = enter(i, source);
-    for (let j = 0; j < datum.length; j++) {
-      target[start + i] = datum[j];
-      i++;
-    }
-  }
-}
-
-/*
- * Use the source data array to fill the target data array
- * @params {TypedArray} source - original data
- * @params {TypedArray} target - output data
- * @params {Number} size - length per datum
- * @params {Function} enter - callback to get new data when source is short
- * @params {Array<Number>} [sourceLayout] - subdivision of the original data in [chunkSize0, chunkSize1, ...]
- * @params {Array<Number>} [targetLayout] - subdivision of the output data in [chunkSize0, chunkSize1, ...]
- */
-export function padArray({source, target, size, enter, sourceLayout, targetLayout}) {
-  if (!Array.isArray(targetLayout)) {
-    // Flat arrays
-    padSubArray({
-      source,
-      target,
-      enter
-    });
-    return target;
-  }
-
-  // Arrays have internal structure
-  let sourceIndex = 0;
-  let targetIndex = 0;
-  const subArrayEnter = (i, context) => enter(i + targetIndex, context);
-
-  const n = Math.min(sourceLayout.length, targetLayout.length);
-
-  for (let i = 0; i < n; i++) {
-    const sourceChunkLength = sourceLayout[i] * size;
-    const targetChunkLength = targetLayout[i] * size;
-
-    padSubArray({
-      source: source.subarray(sourceIndex, sourceIndex + sourceChunkLength),
-      target,
-      start: targetIndex,
-      end: targetIndex + targetChunkLength,
-      enter: subArrayEnter
-    });
-
-    sourceIndex += sourceChunkLength;
-    targetIndex += targetChunkLength;
-  }
-
-  if (targetIndex < target.length) {
-    padSubArray({
-      source: [],
-      target,
-      start: targetIndex,
-      enter: subArrayEnter
-    });
-  }
-
-  return target;
-}
