+++ test/modules/core/utils/index.js
@@ -23,5 +23,6 @@ import './get.spec';
 import './flatten.spec';
 import './positions.spec';
 import './memoize.spec';
+import './array-utils.spec';
 import './scale-utils.spec';
 // import './compare-objects.spec';
